backend/requirements.txt
Flask
paho-mqtt
mysql-connector-python
pycryptodome
flask-cors
python-dotenv